linkstorage/requirements.txt
# requirements.txt 
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
passlib[bcrypt]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
pytest==8.3.5
httpx==0.28.1
email-validator==2.2.0
python-multipart==0.0.20